refactor(android): split background controller concerns
